added loginRepository interface and set interface to view models
